Add files via upload
              Nowadays Educational institutions are concerned about regularity of student attendance. This  is mainly  due to students’  overall academic  performance is affected by his or her attendance in the institute. Mainly there are two conventional methods of marking attendance which are calling out the roll call or by taking student sign on paper. They both were more time consuming and difficult. Hence, there is a requirement of computer-based student attendance management system which will assist the faculty for maintaining attendance record automatically but if we do using mobile phone or not attend class mentally means mark will be reduced.

In this project we have implemented the automated attendance system using Machine Learning. We have projected our ideas to implement “Automated Attendance System Based on Facial Recognition”, in which it imbibes  large  applications.  The application includes face identification, which saves time and eliminates chances of proxy attendance because of the face authorization. Hence, this system can be implemented in a field where attendance plays an important role